Install dependencies & updates

+++ src/Controller/TaskController.php
@@ -7,7 +7,9 @@ use App\Form\TaskType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/tasks')]
 class TaskController extends AbstractController {
@@ -19,14 +21,24 @@ class TaskController extends AbstractController {
     }
 
     #[Route('', name:"task_list", methods: ['GET'])]
-    public function listTask() {
+    #[IsGranted('ROLE_USER')]
+    public function listTask(): Response {
 
         return $this->render('task/list.html.twig',
             ['tasks' => $this->entityManager->getRepository(Task::class)->findAll()]);
     }
 
+    #[Route('/over', name:"task_over", methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
+    public function listTaskOver(): Response {
+
+        return $this->render('task/list-over.html.twig',
+            ['tasks' => $this->entityManager->getRepository(Task::class)->findBy(['isDone' => 1])]);
+    }
+
     #[Route('/create', name:"task_create", methods: ['GET', 'POST'])]
-    public function createTask(Request $request) {
+    #[IsGranted('ROLE_USER')]
+    public function createTask(Request $request): Response {
 
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
@@ -35,6 +47,7 @@ class TaskController extends AbstractController {
 
         if ($form->isSubmitted() && $form->isValid()) {
 
+            $task->setUser($this->getUser());
             $this->entityManager->persist($task);
             $this->entityManager->flush();
 
@@ -47,27 +60,38 @@ class TaskController extends AbstractController {
     }
 
     #[Route('/{id}/edit', name:"task_edit", methods: ['GET', 'POST'])]
-    public function editTask(Task $task, Request $request) {
+    #[IsGranted('ROLE_USER')]
+    public function editTask(Task $task, Request $request): Response {
 
-        $form = $this->createForm(TaskType::class, $task);
-        $form->handleRequest($request);
+        if (in_array('ROLE_ADMIN',$this->getUser()->getRoles()) || $this->getUser() === $task->getUser()) {
 
-        if ($form->isSubmitted() && $form->isValid()) {
+            $form = $this->createForm(TaskType::class, $task);
+            $form->handleRequest($request);
 
-            $this->entityManager->flush();
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+            if ($form->isSubmitted() && $form->isValid()) {
 
-            return $this->redirectToRoute('task_list');
+                $this->entityManager->flush();
+                $this->addFlash('success', 'La tâche a bien été modifiée.');
+
+                return $this->redirectToRoute('task_list');
+            }
+
+            return $this->render('task/edit.html.twig', [
+                'form' => $form->createView(),
+                'task' => $task,
+            ]);
         }
+        throw $this->createAccessDeniedException();
 
-        return $this->render('task/edit.html.twig', [
-            'form' => $form->createView(),
-            'task' => $task,
-        ]);
     }
 
     #[Route('/{id}/toggle', name:"task_toggle", methods: ['GET', 'POST'])]
-    public function toggleTask(Task $task) {
+    #[IsGranted('ROLE_USER')]
+    public function toggleTask(Task $task): Response {
+
+        if ($this->getUser() !== $task->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
 
         $task->toggle(!$task->isDone());
         $this->entityManager->flush();
@@ -78,7 +102,11 @@ class TaskController extends AbstractController {
     }
 
     #[Route('/{id}/delete', name:"task_delete", methods: ['GET', 'POST'])]
-    public function deleteTaskAction(Task $task) {
+    public function deleteTaskAction(Task $task): Response {
+
+        if ($this->getUser() !== $task->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
 
         $this->entityManager->remove($task);
         $this->entityManager->flush();
